Blocks are converted JSON>>HTML
No need for scratchblocks2.js anymore! Still need the stylesheet for
rendering, though. This method allows us to remove the "renderLabel"
and "parameters" in the block data. (The real reason was that we needed
more control to all the block's HTML, mainly for custom blocks. We
don't need to do the algorithm for converting a block label into a spec
anymore, which can't work with the define block).